+++ Sources/SolanaSwift/Programs/TokenPrograms/Token2022Program/InterestBearingConfigToken2022.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public struct ParsedInterestBearingConfig: Codable {
+    static let name: String = "interestBearingConfig"
+    
+    public let currentRate: Int64
+    public let preUpdateAverageRate: Int64
+    public let initializationTimestamp: Int64
+    public let lastUpdateTimestamp: Int64
+    public let rateAuthority: PublicKey
+}
+
+